New Update
This new version includes Stability Updates, But Fixes, a new Icon Class with a Property Editor and a few new controls.

These include a Popup Menu and a blur control

The popup menu's items can currently only be edited in code as I have had a few issues implementing a property editor.